Conflicts:
	CMakeLists.txt.user
	src/qlist_AST.cpp

+++ src/block.cpp
@@ -13,6 +13,7 @@
 #include "tree_element.h"
 #include "document_scene.h"
 
+#include <QDebug>
 /**
  * Block class contructor, that creates a block from a specified element
  * as a child of a specified parent to the specified block group
@@ -90,21 +91,26 @@ Block::Block(TreeElement *el, Block *parentBlock, BlockGroup *blockGroup)
         myTextItem = 0;
         setToolTip(element->getType().replace("_", " "));
 
-        // zistime pocet deti tabulky, tolkokrat spravime next();
-
-        for (int i = 0; i < element->childCount(); i++)
+        int child_count = element->childCount();
+        QList<TreeElement*> children = element->getChildren();
+       // for (int i = 0; i < child_count; i++)children.size()
+       for (int i = 0; i < children.size(); i++)
         {
-            TreeElement *childEl = element->getChildren()[i];
-
+            qDebug() << "child_count " << child_count <<"  size() "<< children.size() << " element:"<< element->getType();
+            TreeElement *childEl = children[i];
             if (!childEl->isFloating()) //! create block from child element
             {
                 new Block(childEl, this);
             }
             else //! create docblock form child element
             {
-                QString text = childEl->getText();
-                childEl->deleteAllChildren();
-                new DocBlock(text, childEl, this, group);
+                if(TreeElement::DYNAMIC){
+                    //! zisti ako to reprezentovat z AST-cka
+                }else{
+                    QString text = childEl->getText();
+                    childEl->deleteAllChildren();
+                    new DocBlock(text, childEl, this, group);
+                }
             }
         }
     }
@@ -399,7 +405,7 @@ Block *Block::removeBlock(bool deleteThis)
     {
         next = 0;
         group->deselect();
-    }    
+    }
 
     foreach (Block* bl, toDelete) //! destroy collected blocks
     {
@@ -1092,7 +1098,7 @@ void Block::updateGeometryAfter(bool doAnimation)
             animate();
 
         group->updateSize(); // root updates group size
-        qDebug("   Geometry updated"); ////////////////////////////
+        qDebug("   Geometry updated");
     }
 }
 
@@ -1137,7 +1143,6 @@ void Block::updatePos(bool updateReal)
                 if (prevSib->showing || lastLeaf->moreSpace)
                 {
                     pos.rx() = prevSib->idealPos().x() + prevSib->idealSize().width();
-
                     //                    offs = prevSib->getOffset(Outer);
                 } //else {
                 //                    offs = lastLeaf->getOffset(Outer);
